More work on hrtim eev
