+++ examples/hrtim_eev.rs
@@ -57,39 +57,29 @@ fn main() -> ! {
 
     let pin_a: PA8<Alternate<AF13>> = gpioa.pa8.into_alternate();
 
-    //        .               .               .  *
-    //        .  33%          .               .  *            .               .
-    //        .-----.         .-----.         .--.            .               .
-    //out1    |     |         |     |         |  |            .               .
-    //        |     |         |     |         |  |            .               .
-    //   ------     -----------     -----------  -----------------------------------
-    //        .               .               .  *            .               .
-    //        .               .               .  *            .               .
-    //        .               .               .  *--------    .               .
-    //fault   .               .               .  |        |   .               .
-    //        .               .               .  |        |   .               .
-    //   -----------------------------------------        --------------------------
-    //        .               .               .  *            .               .
-    //        .               .               .  *            .               .
+    //        .               .  *            .
+    //        .  33%          .  *            .               .               .
+    //        .-----.         .--*            .-----.         .-----.         .-----
+    //out1    |     |         |  |            |     |         |     |         |
+    //        |     |         |  *            |     |         |     |         |
+    //   ------     -----------  --------------     -----------     -----------
+    //        .               .  *            .               .               .
+    //        .               .  *            .               .               .
+    //        .               .  *--------*   .               .               .
+    //eev     .               .  |        |   .               .               .
+    //        .               .  |        |   .               .               .
+    //   -------------------------        ------------------------------------------
+    //        .               .  *            .               .               .
+    //        .               .  *            .               .               .
     let (timer, (mut cr1, _cr2, _cr3, _cr4), mut out1) = dp
         .HRTIM_TIMA
         .pwm_advanced(pin_a, &mut rcc)
         .prescaler(prescaler)
         .period(0xFFFF)
-        //.with_fault_source(fault_source1)
-        //.with_fault_source(fault_source2)
-        .with_fault_source(fault_source3) // Set fault source
-        //.with_fault_source(fault_source4)
-        //.with_fault_source(fault_source5)
-        //.with_fault_source(fault_source6)
-        .fault_action1(FaultAction::ForceInactive)
-        .fault_action2(FaultAction::ForceInactive)
-        // alternated every period with one being
-        // inactive and the other getting to output its wave form
-        // as normal
         .finalize(&mut fault_control);
 
     out1.enable_rst_event(EventSource::Cr1); // Set low on compare match with cr1
+    out1.enable_rst_event(&eev_input3);
     out1.enable_set_event(EventSource::Period); // Set high at new period
     cr1.set_duty(timer.get_period() / 3);
     //unsafe {((HRTIM_COMMON::ptr() as *mut u8).offset(0x14) as *mut u32).write_volatile(1); }
